Modified deleteSwimmer method to manage when the swimmer doesn't exist

+++ com.corbinelli.lorenzo.swimming/src/main/java/com/corbinelli/lorenzo/swimming/controller/SwimmingController.java
@@ -27,6 +27,10 @@ public class SwimmingController {
 	}
 
 	public void deleteSwimmer(Swimmer swimmer) {
+		if(swimmerRepository.findById(swimmer.getId()) == null) {
+			swimmerView.showError("No existing swimmer with id " + swimmer.getId(), swimmer);
+			return;
+		}
 		swimmerRepository.delete(swimmer.getId());
 		swimmerView.swimmerRemoved(swimmer);
 	}
